Init app with custom initAppFromScene()
This adds and uses an alternative approach to
initializing the app provided by @gavrichards:
- Do not call RCTAppDelegate's
application:didFinishLaunchingWithOptions
but instead cherry-pick the code from RCTAppDelegate's
application:didFinishLaunchingWithOptions
except for window and rootViewController creation
- move window and rootViewController creation to
PhoneScene since they're not needed in stand-alone
CarScene
- call initAppFromScene() both in PhoneScene and
CarScene to init the app.

This approach works for both startup scenarios,
both on Phone and on CarPlay-client.

Bonus:
The rootView property is stored in AppDelegate,
so it can be used in PhoneScene
(i.e. to pass to RNBootSplash if used)

Caveat:
The code in initAppFromScene() needs to
be adjusted to the specific version of react native
you are using!
The version used in stand-alone-example is
currently 0.71.13, so the code is taken from that
versions RCTAppDelegate and converted to Swift
(with a little help from ChatGPT for the C++ block,
so no guarantee that it works! I'm not on RN's
new architecture yet).